requirements.txt
numpy~=1.19.1
tensorflow-gpu~=1.15.3
